requirements.txt
opencv-python
Pillow
imagehash
PySide6
numpy
